requirements.txt
aerich==0.7.2
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
bcrypt==4.1.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.0
dictdiffer==0.9.0
dnspython==2.5.0
email-validator==2.0.0.post2
exceptiongroup==1.2.0
fastapi==0.109.0
fastapi-users==12.1.2
fastapi-users-tortoise==0.2.0
h11==0.14.0
idna==3.6
iso8601==1.1.0
makefun==1.15.2
passlib==1.7.4
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pypika-tortoise==0.1.6
python-dotenv==1.0.1
python-multipart==0.0.6
pytz==2023.3.post1
sniffio==1.3.0
starlette==0.35.1
tomlkit==0.12.3
tortoise-orm==0.20.0
typing_extensions==4.9.0
uvicorn==0.27.0
